+++ starlite_users/service.py
@@ -421,87 +421,4 @@ class UserService(Generic[UserModelType, UserCreateDTOType, UserUpdateDTOType]):
         return token
 
 
-def get_session_retrieve_user_handler(user_model: Type[UserModelType], role_model: Type[RoleModelType]) -> Callable:
-    """Factory to get retrieve_user_handler functions for session backends.
-
-    Args:
-        user_model: A subclass of a `User` ORM model.
-    """
-
-    async def retrieve_user_handler(session: Dict[str, Any], connection: ASGIConnection) -> Optional[user_model]:
-        """Handler to register with a Starlite auth backend, specific to SQLAlchemy.
-
-        Args:
-            session: Starlite session
-        """
-        async_session_maker = connection.app.state.session_maker_class
-
-        async with async_session_maker() as async_session:
-            async with async_session.begin():
-                repository = SQLAlchemyUserRepository(
-                    session=async_session, user_model_type=user_model, role_model_type=role_model
-                )
-                try:
-                    user = await repository.get(session.get("user_id", ""))
-                    if user.is_active and user.is_verified:
-                        return user
-                except RepositoryNotFoundException:
-                    return None
-
-    return retrieve_user_handler
-
-
-def get_jwt_retrieve_user_handler(user_model: Type[UserModelType], role_model: Type[RoleModelType]) -> Callable:
-    """Factory to get retrieve_user_handler functions for jwt backends.
-
-    Args:
-        user_model: A subclass of a `User` ORM model.
-    """
-
-    async def retrieve_user_handler(token: Token, connection: ASGIConnection[Any, Any, Any]) -> Optional[user_model]:
-        """Handler to register with a Starlite auth backend, specific to SQLAlchemy.
-
-        Args:
-            token: Encoded JWT
-        """
-        async_session_maker = connection.app.state.session_maker_class
-
-        async with async_session_maker() as async_session:
-            async with async_session.begin():
-                repository = SQLAlchemyUserRepository(
-                    session=async_session, user_model_type=user_model, role_model_type=role_model
-                )
-                try:
-                    user = await repository.get(token.sub)
-                    if user.is_active and user.is_verified:
-                        return user
-                except RepositoryNotFoundException:
-                    return None
-
-    return retrieve_user_handler
-
-
 UserServiceType = TypeVar("UserServiceType", bound=UserService)
-
-
-def get_service_dependency(
-    user_model: Type[UserModelType], role_model: Type[RoleModelType], user_service_class: Type[UserServiceType]
-):
-    """Factory to get service dependencies.
-
-    Args:
-        user_model: A subclass of a `User` ORM model.
-        user_service_class: A subclass of [UserService][starlite_users.service.UserService]
-    """
-
-    def get_service(session: AsyncSession) -> UserServiceType:
-        """Instantiate service and repository for use with DI.
-
-        Args:
-            session: SQLAlchemy AsyncSession
-        """
-        return user_service_class(
-            SQLAlchemyUserRepository(session=session, user_model_type=user_model, role_model_type=role_model)
-        )
-
-    return get_service
